Add .gitignore file
The .gitignore file tells git to ignore files that are routinely
generated. I used gitignore.io to populate the file. I also added
[Ll]ocal to the virtualenv change because at the time of this commit,
the github team has failed to merge my pull request to add this to the
list of virtualenv directories to ignore. If we find more files that
need to be ignored, we should generate the file again and add [Ll]ocal
in if needed. Files for Python, Vim, Pycharm, and VirtualEnv are
ignored.